fix: watch mode tweaks
