+++ .changeset/gentle-deers-sit.md
@@ -1,0 +1,5 @@
+---
+'@hey-api/openapi-ts': patch
+---
+
+fix: watch mode handles servers not exposing HEAD method for spec
