+++ .changeset/happy-eels-kick.md
@@ -1,0 +1,5 @@
+---
+'@hey-api/openapi-ts': patch
+---
+
+fix: add watch.timeout option
