+++ packages/openapi-ts/src/generate/__tests__/class.test.ts
@@ -60,7 +60,8 @@ describe('generateLegacyClientClass', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
