+++ packages/openapi-ts/src/generate/__tests__/core.test.ts
@@ -75,7 +75,8 @@ describe('generateLegacyCore', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -166,7 +167,8 @@ describe('generateLegacyCore', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -240,7 +242,8 @@ describe('generateLegacyCore', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
