+++ packages/openapi-ts/src/generate/__tests__/output.test.ts
@@ -61,7 +61,8 @@ describe('generateLegacyOutput', () => {
       useOptions: false,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
