+++ packages/openapi-ts/src/plugins/@hey-api/schemas/__tests__/schemas.test.ts
@@ -61,7 +61,8 @@ describe('generateLegacySchemas', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -149,7 +150,8 @@ describe('generateLegacySchemas', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
