+++ packages/openapi-ts/src/plugins/@hey-api/sdk/__tests__/plugin.test.ts
@@ -63,7 +63,8 @@ describe('handlerLegacy', () => {
       useOptions: false,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -222,7 +223,8 @@ describe('methodNameBuilder', () => {
       useOptions: false,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -304,7 +306,8 @@ describe('methodNameBuilder', () => {
       useOptions: false,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
 
@@ -388,7 +391,8 @@ describe('methodNameBuilder', () => {
       useOptions: false,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
