+++ packages/openapi-ts/src/plugins/@hey-api/typescript/__tests__/plugin.test.ts
@@ -61,7 +61,8 @@ describe('generateLegacyTypes', () => {
       useOptions: true,
       watch: {
         enabled: false,
-        interval: 1000,
+        interval: 1_000,
+        timeout: 60_000,
       },
     });
