+++ packages/openapi-ts/src/types/client.d.ts
