+++ packages/openapi-ts/src/types/utils.d.ts
