+++ packages/openapi-ts/src/utils/__tests__/parse.test.ts
@@ -38,7 +38,8 @@ describe('operationNameFn', () => {
     useOptions: false,
     watch: {
       enabled: false,
-      interval: 1000,
+      interval: 1_000,
+      timeout: 60_000,
     },
   };
