+++ packages/openapi-ts/test/openapi-ts.config.ts
@@ -6,7 +6,8 @@ export default defineConfig({
     // exclude: '^#/components/schemas/ModelWithCircularReference$',
     // include:
     //   '^(#/components/schemas/import|#/paths/api/v{api-version}/simple/options)$',
-    path: './packages/openapi-ts/test/spec/3.1.x/full.json',
+    // path: './packages/openapi-ts/test/spec/3.1.x/full.json',
+    path: 'http://localhost:8000/openapi.json',
     // path: './test/spec/v3-transforms.json',
     // path: 'https://mongodb-mms-prod-build-server.s3.amazonaws.com/openapi/2caffd88277a4e27c95dcefc7e3b6a63a3b03297-v2-2023-11-15.json',
     // path: 'https://raw.githubusercontent.com/swagger-api/swagger-petstore/master/src/main/resources/openapi.yaml',
@@ -81,8 +82,9 @@ export default defineConfig({
     },
   ],
   // useOptions: false,
-  // watch: {
-  //   enabled: true,
-  //   interval: 1_000,
-  // },
+  watch: {
+    enabled: true,
+    interval: 1_000,
+    timeout: 60_000,
+  },
 });
